pyproject.toml
[project]
name = "imgview"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pillow>=12.1.1",
    "rich>=14.3.3",
]

[project.scripts]
imgview = "main:main"
